refactor: AttendanceTop → AttendanceHeader 리네이밍
* useBlocker 로직 제거
